Add multiple_files plugin
Currently this plugin only gives the ability to pass an array of IOs to
Uploadie#upload. But later users will be able to choose how they want to
do the uploading, and will be able to parallelize them. Uploadie will
also later be able to provide a "parallel" plugin which does basic
parallelization via threads/processes.