Add versions plugin
This plugin enables passing a hash of file versions, getting the same
hash back, only with the files uploaded and files being replaced with
Uploadie::UploadedFile. Currently this exists as a meta plugin for
upcoming "processing" plugin, but it could be later useful with the
addition of "default_location" plugin, where you will be able to specify
the path based on the version name.

If this plugin later turns out not to be useful enough, it will probably
be merged into the "processing" plugin. Same with "multiple_files"
plugin.