Pass through a generic context on uploads
Before this change we were able to pass through a type of the image, and
occasionally a version name. So far this additional information was used
only for generating locations. However, with the addition of the
upcoming Attachable module, users will also want to know the record for
which this location is being generated (to make some conditionals). In
the future, they may be other things as well.

Furthermore, users may also want to have this context during processing
and upcoming validations. We want to give the users complete
flexibility, so we instead have a generic context hash which is passed
everywhere where it needs to.

It may seem simpler that the uploader is simply initialized with that
context. However, we must be careful, we originally wanted the uploader
instance just to be initialized with the storage. If we introduced
initializing with context, we wouldn't be able to generate multiple
versions, because versions are generated already after the uploader was
initialized, so we would have to make an instance for each version we
want to upload, which sounds like it could cause a lot of trouble in the
future. So far this is completely manageable, we'll see how it goes.

One feature that we removed now is that you can't pass anymore a
concrete location where to upload to. But this is actually a good thing,
because the goal of this library is to generate these locations itself.
Also, when doing uploading which will generate versions, it's a bit
strange to be able to pass the location, since it will be completely
ignored anyway.