Put storing metadata into core
This is really nice to have out of the box, users may think they don't
need it, until they do and then it could be too late. This means that
there is no weird internal _metadata plugin, everything is available to
other plugins. And the store_dimensions plugin is now nicely emphasized.
Furthermore, this means that the upcoming validation_helpers plugin can
use this feature.

On the practical side, now we can do validations on standardized cached
files, rather than on whatever gets assigned to the attacher. This means
that we can always cache the file, and not have to worry what to do if
the validations didn't pass. This greatly simplifies the implementation
the attacher would have to have.