Require user-specified libraries in plugins
It's much more convenient to have the library required by itself,
because it's strange to have to require a library yourself, and then you
actually don't see where it's used in the file (because it's actually
used under-the-hood). First I thought thread-safety issues are a
problem, but they're not, people are doing on-the-fly require in Rails
all the time, and they can simply eager load if they need to.