Enable passing in custom extractors
We remove the cumbersome checks as well, because if a user doesn't
follow the contract, in any case they will get a verbose error from
which they can figure out the mistake. It's better not to have all these
duplications and checks, Ruby is a dynamic language and we should treat
it that way.