Extract generic complicated IO stuff into Utils
The Utils module isn't part of the core, and isn't overridable. It
provides a helper for downloading files, which is essentially calling
open-uri, with a patch to always return a Tempfile. It also provides a
helper for creating a temporary copy of an IO.

This module can be used in storages, and it will be used in the
remote_url plugin.