Run validations on attachment assignment
The remote_url validation errors had a bug where if a file was
previously set, for any new remote_url that was invalid download errors
wouldn't be set. The best way to get around that is to do validation
immediately on assignment. This is probably a good decision, for example
it will allow hard validations which raise errors on POROs.