Require version names to be specified
Previously we were exposed to a DoS vulnerability, where attackers would
be able to simply duplicate cache versions infinite number of times, and
then the server would have to upload all of those to store. Instead we
should require users to specify versions that they want to have, and
when generating versions we'll assert if they're correct, but during
attacher assignment we'll just filter out the bad ones, so that
attackers cannot cause errors to the application.