Add :return_url options to endpoint
Often after uploading images to enpdoint AJAX, users will want to
display the uploaded images. So we add the :return_url option to return
the URL to the client. The reason this is not the default is because
there are some storages where generating an URL requires an HTTP
request. Instead of users being baffled in these situations why are
things so slow, we let them consciously make the decision if they want
URLs generated.