Use symbols for version keys instead of strings
We switched to strings only because of security paranoia. But actually
the only vulnerable entry point is Attacher#set. After we've filtered
that, there is no reason to worry and we can switch to symbols again.
Symbols are much more natural to use, and won't confuse the users, since
they likely used symbols in the processing block.