Switch completely to symbols (again)
It was a wrong decision to switch to strings, this was just paranoia for
the Symbols DoS. We now took care of the user input, and in the rest of
the code we use symbols, because it's much more intuitive.

We also remove unnecessary storage existence checks in plugins, there is
very small chance that there will be a typo, and if it happens they will
get a runtime error (they don't have to get the error immediately). This
is just to avoid a lot of boilerplate.