Remove UploadedFile#path
It was meant for less copying during processing, but it's never a good
idea to modify the cached file. For example, if this happens during
promoting in background, if uploading to store lasts longer, the user
may experience a modified or missing image for a short time, which is
definitely what we want to avoid. Users always have to `#download` from
storage if they want to get a file, that's a rule.