Return url inside uploaded file in endpoint
We put it outside so that it isn't saved. But then even if you're not
using :return_url you would have some work on the JavaScript side to
translate the response into what Shrine will accept. This isn't very
friendly, so we'll just put it back inside, and if users mind that it's
saved in the database they can just pull it out with JavaScript. That
makes much more sense