Remove :return_url option from endpoint
When uploading to cache, the URL can easily be deducted from the ID
client side. They're a little harder from URLs like S3, but it's a much
less likely scenario that it will be used with endpoint. We'll add it if
there's a feature request, but for now we can remove the serialization
code.