Remove the rack_file plugin
It's causing too many problems, we always have to check if a Hash is a
Rack File or versions, and in the future there is a great possibility
that we'll forget one time, and get obscure errors. When you're in Rack
app, you simply have to create an IO out of it. You could use the
Endpoint's one:

```rb
file = Shrine::Plugins::Endpoint::RackFile.new(hash)
```