Rebrand extract_content_type to real_content_type
We want to communicate to the users that content type detection provided
by base is not the real one, it's only inferred from the file extension.
So we make it possible for users to detect real content type with
various gems (mimemagic added), but the preferred one is with the `file`
command.

We remove mime-types because it's not part of it, and if people really
want to use it they can just implement it themselves (because then it's
a very specific use case).