Improve cleaning garbage and fix some bugs
Cached files are now automatically deleted when promoted, and keep_files
plugin receives a new :cached option to disable deleting cached files.
We also fix moving plugin to work correctly with promoting, so that
cached files are not attempted to be deleted twice.

We also make keep_files work with background_delete plugin, since in
that case users will call `delete` directly on the uploader.