UploadedFile#content_type > UploadedFile#mime_type
"Content type" referes to the "Content-Type" header, which is whatever
the browser sends in headers. This is different from "MIME type", which
is associated to actual files. We want to make this difference. The MIME
type is by default taken from the Content-Type header, but it can also
be taken from the actual contents with tools like `file`. We can be sure
that this is the right terminology because the `file` UNIX utility uses
the same one.