Rename endpoint plugin to "direct_upload"
"Endpoint" is too generic, from the name of the plugin you can't tell
what it does, just that it provides some kind of endpoint. We want our
plugin names to be memorable, so we want to say exactly what it does.

Also, later we may want to add a plugin for on-the-fly processing, which
will also contain an endpoint, but a different kind.