Allow storages to support multiple deletes
Amazon S3 supports multiple deletes in a single HTTP request, so caring
about performance as we do, we want to leverage that functionality if
storage supports it.

We also add the support for passing an array of files to Shrine.delete,
so that users can leverage it in their custom scripts.