Remove preserve_filename plugin
I don't think it's really necessary, because I can't think of a case
where the application would need to provide downloadable files with the
same filename. This also means that we have to ensure uniqueness of
paths (which I think we did, but I'm not sure), and sanitization of
filenames (which kind of defeats the purpose of "same filename"). If
there is a need for it, we can bring this plugin back, but for now there
is no need to maintain extra code.