Fix resubmitting same file detected as replacement
When the form is saved, and submitted again without file changing, it
will be sent again from the "hidden" field. This would then be detected
as replacement and the file successfully assigned, but it would get
deleted. We change the behaviour to compare the given file and one from
column, and only detect that it's changed when they are indeed
different.