Improve Shrine.uploaded_file
It should also accept UploadedFiles, in that case it just yields. This
is useful when in future we might know in what form is some uploaded
file. Also, it's useful when you want to iterate over files, but you
don't know if it's one or more.

We also add the method to Attacher which delegates to Shrine. This is
useful in backgrounding promoting, so that you don't have to pass the
Shrine class.