Move validate block to the Attacher
Since the validation block is executed in context of the attacher, it's
much more natural to call it on the Attacher class. This also means that
we don't have to worry about which data are we sending through the
arguments, they're right there in the instance so they have access to
anything.

This will also be consistent with the future background plugin where it
will expose much nicer interface of class level blocks, all on the
Attacher as well.