Reuse the existing attachment if it hasn't changed
Attackers can at the moment tamper with the file metadata in the hidden
field, and the current solution is the restore_metadata plugin. However,
restore_metadata always redownloads the file so that it can extract its
metadata again, which is obviously not feasible for any storage that's
not filesystem.

However, there is a much simpler way we can guard ourselves. If we
detect that the file hasn't changed in ID or store, we can simply use
the already written one instead of accepting a potentially tampered.