Don't allow assigning stored files
Until now it was a feature that you can directly upload stored files and
assign them as attachments (along with cached files). This was useful if
people simply wanted to avoid caching.

However, this creates a huge security issue, where attackers can easily
assign themselves someone else's stored files, by reading the location
from the URL. If they can assign foreign attachments, with the default
behaviour it generally means they can easily delete them as well, which
means that the files will be deleted for the other user as well. This
can be avoided by using `:keep_files`, but Shrine should be secure by
default.

One idea that I had was to create signatures for uploaded files, which
would be included in the data hash. Shrine would forbid uploading
attachments which don't have a correct signature (which is calculated by
mixing secret key and file's data hash). This would also trivially solve
tampering of metadata. However, this also isn't secure, because if an
attacker would've broken into the database, they could see the
signatures and delete all the files by assigning them to a record and
destroying the record.

I liked this feature of Shrine, but it isn't worth it if it's insecure,
and I don't think people will miss it much.