Explicitly convert quantity metadata into integers
If the metadata is passed as URL-encoded, either as part of the URL or
in a body of a POST request, all values will be decoded strings, even
numbers.  This produces really hard to catch bugs. Instead we should do
users a favour and automatically convert those strings into integers.