Fix background_helpers and make it do all the work
Firstly we fix background_helpers by putting promoting and deleting
after commit, so that we utilize consistent behaviour. For promoting
this was more necessary, since the background job could have triggered
before the record has managed to save itself, which may cause the job
not to find the record.

Promoting and deleting isn't symmetrical, because promoting uses the
attacher directly and the record needs to exist, while deleting uses
Shrine directly and the record doesn't need to exist (and won't if the
record is destroyed). However, the user of Shrine shouldn't have to know
about that, so instead this commit makes `background_helpers` do all the
work of that the background job has to do.