Allow dimensions to be missing
If there are live uploads in production, previously it could cause
problems to add the store_dimensions, because cached files that weren't
promoted would suddenly raise an error during promotion, because they
don't have the "width" and "height" attributes. Therefore we remove this
constraint and allow both attributes to be missing or nil.