Make file utility able to accept non-files
File utility can also accept the file contents on standard input. The
reason why I didn't add it in the past is because I thought I would have
to read the whole file. But the magic header is actually only in the
beginning of the file, so we just have to read that part.