Remove test with ImageProcessing tempfile on S3
Tempfiles from ImageProcessing previously weren't uploadable to S3 due
to some obscure error, so we added a test here just in case. But I think
it's not needed anymore, because the core behaviour of ImageProcessing
certainly won't change, so there is no way of breaking this test. This
means that we can remove the "ImageMagick" dependency.