Fix infinite loop with recache plugin
Previously loading the recache plugin would cause an infinite loop while
saving the record, because promoting would save the record again, which
would re-trigger recaching, since it didn't check whether the current
file is cached or stored. We now make the recache plugin trigger only on
cached attachment, and we refactor the attacher to make sure it's
triggered only on initial assignment.

The attacher refactoring also made sure that both Attacher#save and
Attacher#finalize will be called only on initial attachment, which makes
things more robust and less likely for future plugins to introduce bugs.

This also fixes another bug when using background jobs, where the
promotion would be called each time the record with a cached attachment
is saved, until the attachment is promoted. This probably wouldn't cause
any errors, but it means the attacker could save a record multiple
times, causing processing and storing to happen multiple times for the
same attachment.

This refactoring introduces backwards compatibility with any existing
external ORM integrations, which will have to update their code.  But
this hopefully ensures that this won't have to happen in the future.