Don't delete cached files
There are couple of problems with deleting cached files. Firstly, when
using backgrounding it can happen that a cached file is deleted (either
by being replaced or destroyed) while being promoted, which would cause
an error. This wouldn't necessarily do any harm, but it would look like
an unnecessary alarm that there is an error, while in reality nothing
erroneous happened.

Secondly, it can cause unexpected errors when using the `moving` plugin,
because we don't know whether the cached file has been moved before
deleting it. This was mitigated by a hack where we make an uploaded file
remember that it was deleted, but that wasn't a good solution, because
we had to modify UploadedFile's data hash. With introduction of this
rule we make sure that these errors cannot happen.

An idea has occured that we can simply rescue any delete errors from the
attacher level, and log a warning. But that seemed wrong, we don't know
what causes these errors, it could be that the underlying storage is
throwing an informative error which we are swallowing. And also this
wouldn't solve other errors which deleting cached files bring, so we can
avoid adding this potentially problematic behaviour.