Remove delete_invalid plugin
The problem with this plugin is that, if the user resubmits an invalid
file, they will get an error because the underlying storage would try to
delete a file that was already deleted. This could cause a lame error in
the production, and we want to follow the principle of least surprise,
so we're removing this plugin, because isn't useful enough to justify
its extistence and these potential errors.