Add support for Sequel's Postgres JSON column
Sequel has a "pg_json" plugin, which serializes and deserializes a
Postgres JSON column into a hash-like object. This object is not a
descendant of Hash, but it responds to #to_hash, so we can loosen the
requirement in Shrine and accept any object that responds to #to_hash as
data.