Bump the "down" dependency to ">= 1.0.3"
The Down gem had two bugs prior to this version. First one was that it
wasn't working with the pretty_location plugin. Second one was that it
wasn't preserving the file extension. So we bump up the required version
so that we ensure that people will use a good version.