Don't require :max_size option in direct_upload
Limiting the maximum file size should ultimately be set on the frontend
server (nginx, apache). This functionality can by bypassed anyway by
uploading the file directly in the form, so this option is only to
fortify the defenses a bit. We will probably make a `constraints` plugin
in the future to ensure no attacks get through.