Remove Shrine.delete
This method is redundant, and requires more code in the base
functionality as well as other plugins. You should always delete files
with a `Shrine` instance, it doesn't matter with which storage it is
initialized. This of breaks backwards compatibility, but hopefully not
many people were using it, and it's better to do it now when releasing
1.0 than to have it exist through 1.x versions.