Enable passing presign options in direct_upload
Often users will want to configure presign options, in the direct_upload
endpoint as well. One natural way seemed to allow passing query
parameters for wanted presign options. But that imposes a security risk
of people being able to configure all presign options (even
permissions), and also it would create a Symbol DoS vulnerability,
because we would have to convert these query parameters to symbols, and
we have no easy way of knowing the list of valid presign options.

So we instead add the ability to configure these options when loading
the plugin. We also yield the request object so that it's possible to
generate different options depending on the request path and parameters.

This introduces a small backwards incompatibility, because we remove
automatic `content_type` assignment via a query parameter, and instead
expect the user to implement that. But I think it really makes sense,
and we're releasing a major version anyway.