Always use non-bang processing methods in docs
The bang methods example were for reducing copying, but it actually
doesn't work as it simply modifies the same image all the time, so you
end up with the same image on all versions. Using non-bang methods has
exactly the required amount of copying, and we now want to recommend not
using bang methods anymore.