Add (before|around|after)_upload to hooks plugin
This is useful mostly for the around_upload hook, if the user wishes to
wrap the whole upload process in a block. At first I thought it's
redundant because `before_upload` is the same as `before_process`, and
`after_upload` is the same as `after_store`, but `around_upload` isn't
redundant, and also it's more readable if the intention revolves around
the whole upload process.