Recommend enabling errors for AR callbacks
ActiveRecord versions 3.x and 4.x automatically silence errors in
callbacks, which makes debugging much more difficult as you don't see
the actual errors if you did a typo or something. So, we recommend to
users that they should enable errors, and this is going to be the
default in ActiveRecord 5 anyway.