Forward `:upload_options` to presign on S3 storage
There are a lot of options which are common for uploading and
presigning, and if an option is not applied on presigning, it won't be
applied when that file is uploaded, since the app won't be used at all
during upload. So we want the options like `:acl` and `:cache_control`
to be applied on both uploading and presigning.

We also extracted uploading and copying into separate methods so that
they can be overridden when the storage is subclassed, for dynamic
options depending on the IO.