Fixes to the previous commit
We can just use `SecureRandom.hex`, which generates a random
alphanumeric with double the length as the given number. We don't need
the test for this behaviour (nothing can break), so we remove it. We
also remove the multiline regex anchors, since there is no risk of
newlines getting in there, and I find single-line anchors much easier to
read.