Make parallelize always work with moving
When the moving plugin does an actual move on the storage, the
parallelize plugin works fine with this. However, when the moving is
done by uploading and deleting, the deleting part isn't processed by the
thread pool. The first problem with this is that it's impacting
performance, but the second actual problem is that it's very probable
that the file will get deleted before it was uploaded.