Don't require that storages rewind the IO
Instead of requiring that every storage rewinds the file, we can make
Shrine itself always rewind the file after uploading. Almost every
uploading will require rewinding, so in most cases we won't do an extra
rewind (if there even is any overhead to it). This way we have rewinding
in a single place.