Make storage linter more flexible
Firstly we add the ability to pass a custom IO factory. This is because
some storages only accept certain types of files (e.g. Cloudinary by
default accepts only images), so we want storage writers to be able to
test with custom files.

Secondly we add the ability to print the lint errors as warnings rather
than exceptions. It can happen that some storages implement eventual
consistency, so e.g. a deleted files can still "exist" immediately after
the deletion. So we want developers to be able to choose lighter way of
informing rather than failure, because they may not be able to do
something about some things.