Close all IOs after uploading them
It's strange that we're not doing this already. We want to always close
the files that we are uploading, so that file descriptors aren't piling
up.

If first thought it was a good idea to rewind files after upload and
keep them open to allow reupload, but the truth is that this use case is
extremely rare. That would mean that people would have to close files
manually. Moreover, it's not always possible to have the former
functionality, because some storages might automatically close the files
(e.g. shrine-cloudinary does it because Cloudinary gem does it
internally). Always closing the file provides consistency and safety.

This imposes a small backwards incompatibility if people were relying in
the file being ready to upload again. But this isn't the case with any
other uploading libraries, nor any http libraries for that matter, so
this shouldn't be expected behaviour anyway.