Use Storage#read in data_uri plugin, add #base64
The most efficient way of retrieving the contents of the file is to use
Storage#read, because firstly this doesn't download the file in a some
storages (a lot of storages alias #open to #download, and we currently
used #open), and secondly this instantiates string on file contents only
once (some storages implement #open by wrapping #read into a StringIO,
and StringIO#read will return a copy of the underlying string, looking
at the Rubinius implementation).

https://github.com/rubysl/rubysl-stringio/blob/da95ae090bc323c3e6e265a97eb2eb272f32d80a/lib/rubysl/stringio/stringio.rb#L431

We also add UploadedFile#base64 which may be convenient, but it's mostly
to refactor the current #data_uri method.