Delete internal Tempfile when closing UploadedFile
If the underlying storage implements `#open` by creating Tempfiles, when
the UploadedFile is closed, we want to delete that internal Tempfile,
because at that point we don't need it anymore. If we don't delete the
Tempfile, it will likely stay laying around, until the process ends
(Tempfiles unfortunately aren't deleted by the GC, only if you use
WeakRef directly). This way we ensured that if a cached UploadedFile is
uploaded to store, any temporary files will get deleted.

One idea I had was to implement `#open` on all storages in terms of
StringIOs, so we wouldn't have to care about deleting. However, that
means that `#open` would load the entire file contents in memory, since
that's how StringIO is internally implemented.