Return extension from filename if ID doesn't have
Some storages generate their own ID on upload (SQL, Flickr, GridFS) in a
way that the extension is lost. In this case we want to try to retrieve
the extension from the filename.

Just to note here that the order of retrieving the extension is
intentional. We first want to try the ID, because some storages like
Cloudinary accept changing format on upload, in which case the ID is
uploaded to the new format, so we want to pick that up. And only
afterwards try the filename.