Bump up "down" dependency to ">= 1.0.5"
The latest version didn't download the file in an optimal way. Down
works in a way that it first downloads the file, then it "fixes" the
extension (which open-uri doesn't add). However, before this version
Down would copy to the new path with extension, and delete the old file
(so that we don't have to make a special StringIO case). This was fixed
and now the file is now moved instead of copied. This means downloads
are now faster, especially for bigger files.