Return the result in around_* hooks
Currently the return value of calling around_* hooks would be whatever
was returned in the corresponding after_* hook. This is not ideal,
because this way it's not possible to get to the information about
generated file locations, extracted metadata, or the result of
processing.

Instead we return the result of the corresponding operation, which will
always return the IO or UploadedFile, which allows hooks to act
according to the result.

class MyUploader < Shrine
  def around_upload(io, context)
    result = super
    Notificatino.send("File was uploaded to location #{result.id}")
    result
  end
end