Small refactoring for simplicity
This does introduce a small backwards incompatibility if people were
relying that :name and :record will be the only two keys in the
`context` hash when overriding `#extract_metadata`, and used the keyword
argument syntax instead of using only the `context` variable. But people
shouldn't have been relying on this, because it would eventually break,
for example when doing file migrations. Also, there is an extremely low
possibility that anyone was relying on this.