Minor README tweaks
The only notable thing was that we removed the "attr_accessor"
declaration. This is because it adds confusion to newcomers that this is
what they should do when using an ORM, while it was in fact only
applicable to ORM-less situations (which are very rare anyway), and
doing it with an ORM causes errors.