Inherit custom metadata from cached files
Previously if custom metadata was assigned to a cached file, it wasn't
forwarded onto the stored file, only the built-in metadata which have
reader methods were inherited (because extraction methods were designed
to detect them). Now if an UploadedFile is given, we directly copy over
its metadata hash, so that all metadata is inherited.

This change also removes the confusion of `#extract_metadata` being
called twice (once for caching and then again for storing). This
behaviour wasn't documented and isn't logical, because we only want to
extract metadata from original files, not from cached files.

Fixes #30