Don't error in backgrounding on missing record
It can happen that a record gets deleted before updating (processing and
storing can take a while, so the user may have already updated the
attachment) or at the beginning of job (either the job started later, or
it's been repeating for some time). In both cases we should handle
missing records in a silent way.

The only argument for keeping these errors would be that we're
potentially suppressing anomalous behaviour. For example, these "record
not found" errors might indicate that (a) jobs are starting longer than
usual, (b) job is erroring, and than the job suddenly completes, or (c)
record finding might not be working properly (e.g. sharding).

However, for none of this cases should Shrine be the indicator; (a) a
generic mechanism should exist for checking if the queue is clogged, (b)
when the job errors, this error should be reported even if the job
retries, and (c) malfunction of sharding will have more global effects
anyway.

One thing to note here is that the reason we're rescuing the generic
`Sequel::Error` instead of `Sequel::NoExistingObject` is to keep
compatibility with Sequel < 2.28, because prior to 2.28
`Sequel::Model#refresh` and `#reload` would raise `Sequel::Error`.