Don't promote if attachment changed in background
It can happen that the queue gets clogged up making the job starts
later, or the job is retrying, so the attachment may have already
changed in the beginning of the job. In this case there is no point in
continuing the processing and uploading, since the promoting will abort
before the update anyway.

This mitigates an unlikely attack where the user can update a corrupted
file for which they predict will break the processing, and they can
continue updating the attachment with the same corrupted file, which can
create a lot of retrying jobs. This probably wouldn't do any harm to the
system, but it's still nice to have that protected, just in case.