Improve aborting update when attachment changed
Previously Shrine checked if attachment has changed outside of a
transaction. However, that means that the attachment can still change
between the check and the update, which could result into new attachment
being replaced with the old one. So we change the code to wrap both the
check and the update in a transaction.