Return the record in backgrounding methods
It's useful if the user wants to do something with the record, that they
can reuse the already instantiated one.

The choice of returning nil when promoting was aborted is deliberate.
We want to ensure that any potential action after promotion isn't
performed, because the promotion in fact didn't happen.