Add the new #update method convention for storages
Some storages have a concept of updating data of already uploaded files:

* Flickr - can update title or description of images
* Cloudinary - can update things like eager transformations
* YouTube - can update title or description of videos

It makes sense to have a convention for that, so we're making it the
`#update` method. This method currently isn't used in any plugin, but it
might in the future. In any case, it's nice to have a convention.