Don't require that metadata values are present
Currently calling `#original_filename`, `#size` or `#mime_type` will
error if the corresponding key-value pair is missing from the metadata
hash. This is not the desired behaviour, because then we can't inquire
whether certain information exists through these methods, we have to
check the metadata hash.

This isn't a problem when the uploaded file is generated through the
Shrine uploader, because there the keys will automatically be present,
even though the values are nil. But this may be a problem when
generating the data hash in a custom way, where we may not have these
metadata available and we don't need them anyway.

The "size" metadata is sometimes required when uploading, it's only not
when uploading to FileSystem, or when copying to S3. But we don't want
to `fetch` it for consistency with other metadata methods, and "size"
being required is already mentioned in the documentation.