Pretty location plugin with namespaced models
When pretty location generates a location for a namespaced model it should replace '::' with '/'. If this is not done the S3 storage's URI method (given host option is set) attempts to use `URI.join` which does not produce an expected result when the path contains '::'.

2.2.3 :001 > require 'uri'
 => true
2.2.3 :002 > URI.join('https://example.com', 'namespaced::openstruct/something.jpg').to_s
 => "namespaced::openstruct/something.jpg"
2.2.3 :003 > URI.join('https://example.com', 'namespaced/openstruct/something.jpg').to_s
 => "https://example.com/namespaced/openstruct/something.jpg"

Besides this one could argue that the directory structure should follow the namespacing of the models.