Add :namespace option to pretty_location
Most often models when models are namespaced, the namespace doesn't hold
a semantical value (e.g. "MyApp::Models::User"). Usually only the model
name is important, since it maps to database tables which are in a
"flat" namespace. For that reason we don't include the namespace by
default.

However, sometimes the namespace can hold a semantical value, so we
provide the option to include the namespace in the location. We also ask
the user to choose the separator for the namesapce, as preferences will
likely differ.

Fixes #47