Give better error messages in remote_url
Currently remote_url by default returns the validation error: "file not
found or is to large". That's not great user experience, that's why we
instead create the validation error from the message of the raised
error. That's why we bump Down to 2.0.1, because this update brings
human-readable error messages.

If the :downloader returns nil, a default "download failed" error
message is used.