Add :include_error option to remote_url
Sometimes we want to generate remote_url messages more accurately, in
which case we need to have the error instance, so that we can generate a
different message depending on what went wrong.

This commit also makes a modification to existing error messages;
instead of

"file not found"
"file is too large (max is 10MB)"

the error messages are now:

"download failed: file not found"
"download failed: file is too large (max is 10MB)"

It think this message is more explicit to the end user. And in case of
some custom error message having obscure meaning, they will always know
that "download failed".