Don't require "metadata" in UploadedFile
In a scenario where the uploaded file data hash is POSTed, and there is
no metadata, URL-encoding doesn't have a specification for parameters
whos values are empty hashes. So when decoding these parameters, the
"metadata" key would be absent, which would cause an error when
initializing an UploadedFile.

Since no metadata is strictly required, it doesn't make sense to demand
the "metadata" key.