Make metadata accessible when generating location
Sometimes we may want generate location differently depending on the
extracted metadata of the uploaded file. For example, when assigning a
data URI, we might want to add an extension depending on the given
content type.

To achieve this, we switch the order, so that metadata is extracted
before generating location. This makes sense, because there cannot be a
scenario where metadata extraction depends on the generated location.

Before this change, it was possible to obtain the metadata by calling
`#extract_metadata` again on the IO, but this is now much faster.