Don't mutate the context inside the uploader
We should generally never mutate input arguments, unless there is a good
reason for it. Here we don't have a good reason, and it could have
caused a bug with repeating attacher assignment (since the context hash
was memoized to an instance variable). Luckily, since we're passing
`:phase` as well, the uploaders received a copy of that hash, so no harm
done. But we see that it could have caused a dangerous bug.