Don't cache data keys in UploadedFile#initialize
Caching into instance variables in this case can create surprising
behaviour. For example, when we want to duplicate an UploadedFile and
change its data (like in the backup plugin), the change won't be picked
up because keys of original data was cached during initialization.

Yes, it does come at a small performance cost of always reading the
hash, but since the data hash will always have three elements, we can
safely discard this concern.

We keep #fetch-ing the three keys from the data hash to ensure that
everything is proper. This change has a small downside that this
assertion of required keys doesn't happen on file initialization
anymore, but invalid data hashes will occur very rarely, and the
appropriate error will be raised eventually.