Make keep_files work better with backgrounding
Since keep_files was overriding `Shrine`, when combined with
backgrounding plugin, the delete background job would always be spawned,
regardless of whether the file will actually get deleted or not.

Moving the plugin to Attacher scope has the advantage that we can get
above `Attacher#delete!` in the call chain, which is what backgrounding
overrides. Additional reason why it shouldn't be overriding
`Shrine#delete` is that `Shrine` should be the "low-level" uploader, so
`Shrine#upload` and `Shrine#delete` should always upload and delete,
regardless of which plugins are loaded.