Add :fallbacks to versions for fallback URLs
When using the recache plugin to generate some versions in the
foreground and some in the background, it is useful to be able to
specify fallback URLs for versions that are processed later, so that
there can be graceful degradation until the background job finishes.

While we can always achieve the same behaviour directly:

  @attacher.url(:thumb_2x) || @attacher.url(:thumb)

it is much nicer to keep the view level unaware of background jobs.