Store backup only after promoting
Currently the backup was done immediately after storing, and only
afterwards promoting was done. This is not ideal, because then updating
the record with processed/stored files is delayed by the duration of
backing up. Since backup can be performed at any point, we move it to
happen after promoting instead.