Make deleting backup work with backgrounding
When using the backup plugin with backgrounding, two scenarios could
happen when deleting a file:

* if backgrounding plugin is loaded *before* backup plugin, deleting the
  backup would happen in the foreground
* if backgrounding plugin is loaded *after* backup plugin, the backup
  wouldn't get deleted at all

So in both scenarios the behaviour was broken. So we move deleting of
backup higher in the call stack, so that it happens "above"
backgrounding (which overrides Attacher#delete!).

In Attacher#replace and Attacher#destroy we rely on the return value to
know whether a stored file was deleted. These methods can return nil if
there is no file or the file is cached, or if keep_files plugin is
loaded. In backgrounding plugin we also explicitly return the stored
file after the job has been kicked off, so that we satisfy this
contract.