Log number of input and output files on processing
The reason why we can't have only number of *output* files is because
the message "3 files processed" indicates that 3 files were processed.
This doesn't have to be true, one could be just an unprocessed original,
or when using the recache plugin only the 3rd version could actually
have been processed.

The reason why we can't have only number of *input* files is because the
message "1 file processed" indicates that there was only a single file
on input and output, so it doesn't differentiate from the case when
we're processing multiple versions.

The only way to display a complete message is to display number of
*both* input and output files.

  PROCESS[store] ImageUploader[:avatar] 1-3 files
  PROCESS[store] ImageUploader[:avatar] 3-5 files

  {"action":"process", "phase":"store", "files":"1-3", ...}

  action=process phase=store files=1-3

This introduces a backwards incompatibility for people who wrote their
own logging format using Shrine#log. But it's highly unlikely that
someone did that.