Add option for non-pollution in migration_helpers
By default migration_helpers will add a lot of methods to the model,
which delegate to the underlying attacher. I thought this was good and
convenient, because I assumed this plugin will only be used in separate
scripts, where the method pollution doesn't matter. But people use it in
their apps, so we add `:delegate` option for people to choose whether
they want to use methods through the model or attacher directly.

While we're here we also improve the documentation of migration_helpers.