Fix S3 URL not being URL-encoded with :host option
If an object key contains spaces or other URL-unsafe characters, and
:host option is being used, generating an S3 URL would fail with a
URI::InvalidURIError, because URI.join internally calls URI.parse, which
expects the given URL to already by URL-encoded.

Instead we can just generate a public URL (because we don't want any of
the signature query parameters, at least for keeping backwards
compatibility), and then take the path part to join with the :host.

Fixes #54