Add tests for promoting when record is invalid
Sometimes users will want to save the record with skipping validations.
Before this didn't work in Shrine, because the first save with cached
file succeeded, but the second save with stored file didn't and would
error, because `record.save!` was used which triggers validations and
errors if record is invalid.

I have accidentally fixed this in 5d7cab6, so now I'm just adding tests
around this so that nobody accidentally breaks it in the future.