Fix path-style URLs not working with :host
Sometimes the S3 URL can be generated like

http://s3.amazonaws.com/bucket-name/file.jpg

instead of

http://bucket-name.s3.amazonaws.com/file.jpg

The former is generated when `:force_path_style` is set to true, but it
can also be generated in other scenarios. In that case the bucket name
is part of the path, and we want to remove it when generating a URL with
:host, because otherwise it's not the correct URL (the CDN host is
linked to bucket-name.s3.amazonaws.com domain, that means that there
cannot be an additional bucket-name in the path).