Add ability to do custom backgrounding
We add `Attacher.dump` and `Attacher.load` for implementing custom
background jobs with custom functionality. `Attacher.dump` dumps the
attacher in a serializable hash suitable for passing as an argument to a
background job, and `Attacher.load` takes that hash and resolves the
attacher from it.

One use case for this is, if the record has a slug which is used for
generating the location of its attachment, the user might want to
reupload the attachment to a new location when slug changes.

  class Record < Sequel::Model
    def after_commit
      if (previous_changes || {}).key?(:slug) && image_attacher.stored?
        ReuploadJob.perform_async(Shrine::Attacher.dump(image_attacher))
      end
    end
  end

  class ReuploadJob
    include Sidekiq::Worker
    def perform(data)
      attacher = Shrine::Attacher.load(data)
      uploaded_file = attacher.uploaded_file(data["uploaded_file"])
      return if uploaded_file != attacher.get

      attacher.promote(uploaded_file)
      attacher.store.delete(uploaded_file)
    end
  end